+++ bastion/resource_externalauth_kerberos.go
@@ -20,6 +20,7 @@ type jsonExternalAuthKerberos struct {
 	Description          string `json:"description"`
 	Host                 string `json:"host"`
 	KerDomController     string `json:"ker_dom_controller"`
+	KeyTab               string `json:"keytab,omitempty"`
 	LoginAttribute       string `json:"login_attribute"`
 	Type                 string `json:"type"`
 }
@@ -57,6 +58,10 @@ func resourceExternalAuthKerberos() *schema.Resource {
 				ForceNew: true,
 			},
 			"description": {
+				Type:     schema.TypeString,
+				Optional: true,
+			},
+			"keytab": {
 				Type:     schema.TypeString,
 				Optional: true,
 			},
@@ -243,6 +248,7 @@ func prepareExternalAuthKerberosJSON(d *schema.ResourceData) jsonExternalAuthKer
 		KerDomController:     d.Get("ker_dom_controller").(string),
 		Port:                 d.Get("port").(int),
 		Description:          d.Get("description").(string),
+		KeyTab:               d.Get("keytab").(string),
 		LoginAttribute:       d.Get("login_attribute").(string),
 		UsePrimaryAuthDomain: d.Get("use_primary_auth_domain").(bool),
 		Type:                 "KERBEROS",