+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 <!-- markdownlint-disable-file MD013 MD041 -->
 # changelog
 
+FEATURES:
+
+* add `wallix-bastion_version` data source
+
 ## 0.5.0 (December 9, 2021)
 
 NOTES: